Version 1.1.7 (#16)
* version 1.1.7

* update

* update

* update

* update

* update

* add media group name

* update media filter

* update media filter

* move isAdminAccessTokenHttpOnly to common

* update access token genrator

* update media repo

* update media repo

* update

* update

* update

+++ ezyplatform-sdk/ezyplatform-common-sdk/src/main/java/org/youngmonkeys/ezyplatform/entity/DataMapping.java
@@ -16,9 +16,17 @@
 
 package org.youngmonkeys.ezyplatform.entity;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.IdClass;
+import javax.persistence.Table;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
@@ -47,7 +55,7 @@ public class DataMapping {
     private long toDataId;
 
     @Column(name = "display_order")
-    private int displayOrder;
+    private long displayOrder;
 
     private BigInteger quantity;
