+++ ezyplatform-sdk/ezyplatform-common-sdk/src/main/java/org/youngmonkeys/ezyplatform/model/DataMappingModel.java
@@ -29,7 +29,7 @@ public class DataMappingModel {
     private String mappingName;
     private long fromDataId;
     private long toDataId;
-    private int displayOrder;
+    private long displayOrder;
     private BigInteger quantity;
     private BigInteger remainingQuantity;
     private BigInteger numberData;
