+++ ezyplatform-sdk/ezyplatform-http-sdk/src/main/java/org/youngmonkeys/ezyplatform/pagination/DisplayOrderAscToDataIdAscDataMappingPaginationParameter.java
@@ -32,7 +32,7 @@ import static org.youngmonkeys.ezyplatform.util.Values.isAllNull;
 public class DisplayOrderAscToDataIdAscDataMappingPaginationParameter
     implements DataMappingPaginationParameter {
 
-    public int displayOrder;
+    public long displayOrder;
     public Long toDataId;
 
     @Override
